Hopefully fixed the UI issue with the Content Copy. TLDR; made content copy modal unable to exit by clicking out
Background - it seems like a lot of people are having issues with the content copy. There have been several tickets which report users thinking that the added a content copy -- but then finding out that they hadn't. I think people just clicking out of the modal and not clicking 'Add Source' which is required to trigger some js which formats the field.